split epoch tests to 2 sub-suites
