+++ integration/tests/epochs/cohort1/epoch_join_and_leave_an_test.go
@@ -1,10 +1,11 @@
-package epochs
+package cohort1
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/suite"
 
+	"github.com/onflow/flow-go/integration/tests/epochs"
 	"github.com/onflow/flow-go/model/flow"
 )
 
@@ -13,7 +14,7 @@ func TestEpochJoinAndLeaveAN(t *testing.T) {
 }
 
 type EpochJoinAndLeaveANSuite struct {
-	DynamicEpochTransitionSuite
+	epochs.DynamicEpochTransitionSuite
 }
 
 // TestEpochJoinAndLeaveAN should update access nodes and assert healthy network conditions
